+++ tests/testapp/models.py
@@ -2,18 +2,12 @@ from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
 from content_editor.models import Region, Template, create_plugin_base
-from cte_forest.models import CTENodeManager
 from feincms3 import plugins
 from feincms3.apps import AppsMixin, reverse_app
 from feincms3.mixins import (
     LanguageMixin, MenuMixin, RedirectMixin, TemplateMixin,
 )
 from feincms3.pages import AbstractPage
-
-
-class PageManager(CTENodeManager):
-    def active(self):
-        return self.filter(is_active=True)
 
 
 class Page(
@@ -70,8 +64,6 @@ class Page(
             'urlconf': 'testapp.articles_urls',
         }),
     ]
-
-    objects = PageManager()
 
 
 PagePlugin = create_plugin_base(Page)